+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 * [#920](https://github.com/clojure-emacs/cider/issues/920): Support `cider-jack-in` for boot-based projects.
 * [#949](https://github.com/clojure-emacs/cider/issues/949): New custom var: `cider-default-repl-command`.
 * New code formatting commands - `cider-format-buffer`, `cider-format-region` and `cider-format-defun`.
+* Pretty printing functionality moved to middleware, adding support for ClojureScript.
+  - New command to eval and pprint result: `cider-interactive-pprint-eval`.
+  - `cider-format-pprint-eval` has been removed.
 
 ### Changes